Split the config file
